Move serialization and deserialization to top-level directory
